requirements.txt
# ============================================================
# fineweb-pipeline requirements.txt
# 预训练数据清洗三代方法论对比实验
# ============================================================

# ── 核心数据处理框架 ──────────────────────────────────────
datatrove[all]>=0.3.0          # HuggingFace 数据处理框架（含 WARC 读取）
datasets>=2.14.0               # HuggingFace Datasets（下载参考数据）
huggingface-hub>=0.20.0        # HF Hub 客户端

# ── 第一代：Heuristic Filtering ───────────────────────────
trafilatura>=1.6.0             # HTML 文本提取
justext>=3.0.0                 # HTML 文本提取（对比实验用）
fasttext-wheel>=0.9.2          # 语言识别 + 质量分类器
langdetect>=1.0.9              # 语言识别备选

# ── 第二代：Model-based Filtering ─────────────────────────
scikit-learn>=1.3.0            # TF-IDF + LogisticRegression 分类器

# ── 毒性检测 ──────────────────────────────────────────────
detoxify>=0.5.0                # 多维度毒性打分
transformers>=4.35.0           # BERT 模型基础

# ── Perplexity 打分 ───────────────────────────────────────
# kenlm 安装较复杂，先用 GPT-2 替代
# kenlm                        # 若需要安装：pip install https://github.com/kpu/kenlm/archive/master.zip

# ── 第三代：API 调用 ──────────────────────────────────────
anthropic>=0.36.0              # Claude API（LLM 改写）
openai>=1.0.0                  # OpenAI API 备选

# ── Proxy Model 训练 ──────────────────────────────────────
torch>=2.1.0                   # PyTorch（支持 MPS backend）
tiktoken>=0.5.0                # OpenAI tokenizer（GPT-2 风格）
lm-eval>=0.4.0                 # lm-evaluation-harness

# ── Notebook 和可视化 ─────────────────────────────────────
jupyter>=1.0.0
ipykernel>=6.0.0
matplotlib>=3.7.0
seaborn>=0.13.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
tqdm>=4.65.0

# ── 工具库 ────────────────────────────────────────────────
pyyaml>=6.0                    # YAML 配置文件读取
requests>=2.31.0               # HTTP 请求（下载数据）
warcio>=1.7.4                  # WARC 文件读取（datatrove 依赖）
xxhash>=3.3.0                  # 高性能哈希（精确去重）
datasketch>=1.6.0              # MinHash 去重
mmh3>=3.1.0                    # MurmurHash3（MinHash 用）

# ── 代码质量 ──────────────────────────────────────────────
black>=23.0.0
flake8>=6.0.0
